Move and fix test test_table_references_in_plan_to_sql
